add SetOperatorsValidationTest
